refactor: move cache selector to internal/managercache package
